Changes with eslint
