+++ app/next.config.js
@@ -3,6 +3,7 @@ import("dotenv").then((o) => o.config());
 module.exports = {
   env: {
     GOOGLE_API_KEY: process.env.GOOGLE_API_KEY,
+    NOTION_API_KEY: process.env.NOTION_API_KEY
   },
   future: {
     webpack5: true,
@@ -11,12 +12,12 @@ module.exports = {
     return [
       {
         source: "/",
-        destination: "/theatre",
+        destination: "/train",
         permanent: false,
       },
       {
         source: "/home",
-        destination: "/theatre",
+        destination: "/train",
         permanent: false,
       },
     ];
