+++ app/src/pages/train/index.tsx
