+++ app/src/pages/url/index.tsx
@@ -1,184 +1,189 @@
-/* eslint-disable */
-import { useRouter } from "next/router";
 import { useEffect, useRef, useState } from "react";
-import { FaExclamationCircle } from "react-icons/fa";
-import firebase from "firebase/app";
-import "firebase/auth";
-import "firebase/firestore";
 
-import style from "./style.module.css";
-import { User, UserRole } from "../../model/user";
+import {
+  Table,
+  Thead,
+  Tbody,
+  Tfoot,
+  Tr,
+  Th,
+  Td,
+  Button,
+  MenuButton,
+  MenuList,
+  MenuItem,
+  Menu,
+  IconButton,
+  Modal,
+  ModalOverlay,
+  ModalContent,
+  ModalCloseButton,
+  ModalBody,
+  useDisclosure,
+  FormControl,
+  Input,
+  FormLabel,
+  ModalHeader,
+  ModalFooter,
+  InputGroup,
+  InputLeftAddon,
+  useToast,
+  Text,
+} from "@chakra-ui/react";
 
-let deepLinkField = "";
-let nameField = "";
-let suffixField = "";
+import { FaEllipsisH, FaPlus, FaTrash } from "react-icons/fa";
+import { provider } from "../../model/provider";
 
-let createdSuffix = "ERROR";
+export default function URLsPage() {
+  const [urls, setUrls] = useState([]);
+  const { isOpen, onOpen, onClose } = useDisclosure();
 
-export default function UrlPage(props: { user: User }) {
-  const router = useRouter();
+  const urlRef = useRef(null);
+  const suffixRef = useRef(null);
 
-  const [adminErrorMsg, setAdminErrorMsg] = useState(``);
-
-  let deepLinkFieldRef = useRef(null);
-  let suffixFieldRef = useRef(null);
-
-  nameField = suffixField = deepLinkField = "";
+  const toast = useToast();
 
   useEffect(() => {
-    if (props.user?.role != UserRole.ExCo) console.log("oop");
-  }, [props.user]);
-
-  function createUrl() {
-    async () => {
-      if (nameField === "" || deepLinkField === "") {
-        setAdminErrorMsg(`Please fill in all required fields.`);
-        return;
-      } else if (
-        (await firebase
-          .firestore()
-          .collection("links")
-          .get()
-          .then((col) => {
-            for (const doc of col.docs) {
-              if (suffixField === doc.data().suffix) {
-                return true;
-              }
-            }
-            return;
-          })) === true
-      ) {
-        setAdminErrorMsg(`This link is already in use.`);
-        return;
-      } else if (
-        deepLinkField.search(
-          /(http|https):\/\/([a-zA-Z0-9][a-zA-Z0-9-]{0,}\.){1,}[a-zA-Z0-9]{2,}/
-        )
-      ) {
-        setAdminErrorMsg(`Invalid deep/long link provided`);
-        return;
-      } else if (
-        deepLinkField.includes(window.location.href) ||
-        deepLinkField.includes("bit.ly") ||
-        deepLinkField.includes("tinyurl.com") ||
-        deepLinkField.includes("goo.gl")
-      ) {
-        setAdminErrorMsg(`Site blacklisted`);
-        return;
-      } else if (
-        deepLinkField.includes("dQw4w9WgXcQ") ||
-        deepLinkField.includes("ub82Xb1C8os") ||
-        deepLinkField.includes("8ybW48rKBME")
-      ) {
-        window.location.href = "https://www.youtube.com/watch?v=dQw4w9WgXcQ";
-        return;
-      }
-
-      const genRandAlias = () => {
-        const characters =
-          "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+-";
-        let randomAlias = "";
-        for (var i = 0; i < 4; i++) {
-          randomAlias += characters.charAt(
-            Math.floor(Math.random() * characters.length)
-          );
-        }
-        firebase
-          .firestore()
-          .collection("links")
-          .get()
-          .then((col) => {
-            col.docs.map((doc) => {
-              if (randomAlias === doc.data().suffix) {
-                return genRandAlias();
-              }
-            });
-          });
-        return randomAlias;
-      };
-
-      if (props.user?.role >= UserRole.ExCo) {
-        const newRandomAlias = genRandAlias();
-        firebase
-          .firestore()
-          .collection(`links`)
-          .doc(suffixField || newRandomAlias)
-          .set({
-            name: nameField ?? "ERROR",
-            link: deepLinkField ?? "ERROR",
-            suffix: suffixField || newRandomAlias,
-            // date: firebase.firestore.Timestamp.fromDate(new Date()),
-          })
-          .then(() => {
-            createdSuffix = suffixField || newRandomAlias;
-            console.log(`Successful document written with ID: ${suffixField}`);
-            router.push(`/success`);
-          })
-          .catch((error) => {
-            console.error(`Error adding document: ${error}`);
-          });
-      }
+    const unsubscribe = provider.url.getURLsSnapshot((snapshot) => {
+      setUrls(
+        snapshot.docs.map((e) => ({
+          url: e.data().url,
+          suffix: e.id,
+        }))
+      );
+    });
+    return () => {
+      unsubscribe();
     };
+  }, []);
+
+  function saveURL() {
+    const valid = validateURL();
+    if (valid !== undefined) {
+      provider.url.updateURL(valid.suffix, valid.url);
+      onClose();
+      toast({ title: "Shortened URL saved successfully!", status: "success" });
+    }
+    toast({ title: "An error occurred while saving", status: "error" });
+  }
+
+  function validateURL(): { suffix: string; url: string } {
+    let url: string = urlRef.current.value;
+    let suffix: string = suffixRef.current.value;
+    if (!(url.startsWith("http://") || url.startsWith("https://"))) {
+      url = `https://${url}`;
+    }
+    if (suffix === "") {
+      suffix = "a"; // random
+    }
+    if (
+      /^https?:\/\/[a-zA-Z0-9-._~!*'();:@&=+$,\/?#\[\]%]+$/g.test(url) &&
+      /^\w+$/g.test(suffix)
+    ) {
+      return { url, suffix };
+    }
+    return undefined;
+  }
+
+  function deleteURL(suffix: string) {
+    provider.url.deleteURL(suffix);
   }
 
   return (
-    <div className={style.main}>
-      <h3>Create a new Link</h3>
-      <div className={style.newFieldDiv}>
-        <LabelField
-          label="NAME*"
-          placeholder="SST Inc Website"
-          onChange={(event) => {
-            nameField = event.target.value;
-          }}
-          onKeyDown={(event) => {
-            if (event.keyCode === 13) {
-              event.preventDefault();
-              deepLinkFieldRef.current.focus();
-            }
-          }}
-        />
-        <LabelField
-          label="LONG URL*"
-          placeholder="https://sstinc.org"
-          onChange={(event) => {
-            deepLinkField = event.target.value;
-          }}
-          onKeyDown={(event) => {
-            if (event.keyCode === 13) {
-              event.preventDefault();
-              suffixFieldRef.current.focus();
-            }
-          }}
-        />
-        <LabelField
-          label="SHORT URL"
-          placeholder="sstinc"
-          onChange={(event) => {
-            suffixField = event.target.value;
-          }}
-          onKeyDown={(event) => {
-            if (event.keyCode === 13) {
-              event.preventDefault();
-              createUrl();
-            }
-          }}
-          leftDetail={<p className={style.domain}>go.sstinc.org/</p>}
-        />
-        <p className={style.smallText}>
-          Leaving the field empty would create a randomly generated alias.
-        </p>
+    <>
+      <Modal initialFocusRef={urlRef} isOpen={isOpen} onClose={onClose}>
+        <ModalOverlay />
+        <ModalContent>
+          <ModalHeader>Create a new shortened URL</ModalHeader>
+          <ModalCloseButton />
+          <ModalBody pb={6}>
+            <FormControl>
+              <FormLabel>Long URL</FormLabel>
+              <Input
+                ref={urlRef}
+                onKeyDown={(e) => {
+                  if (e.key === "Enter" || e.keyCode === 13) {
+                    suffixRef.current.focus();
+                  }
+                }}
+                placeholder="sstinc.org"
+              />
+            </FormControl>
+
+            <FormControl mt={4}>
+              <FormLabel>Short URL (Suffix)</FormLabel>
+              <InputGroup>
+                <InputLeftAddon>smp.sstinc.org/</InputLeftAddon>
+                <Input
+                  ref={suffixRef}
+                  onKeyDown={(e) => {
+                    if (e.key === "Enter" || e.keyCode === 13) saveURL();
+                  }}
+                  placeholder="sstinc"
+                />
+              </InputGroup>
+              <Text size="xs">Leaving the field empty would create a randomly generated alias.</Text>
+            </FormControl>
+          </ModalBody>
+
+          <ModalFooter>
+            <Button colorScheme="blue" mr={3} onClick={saveURL}>
+              Save
+            </Button>
+            <Button onClick={onClose}>Cancel</Button>
+          </ModalFooter>
+        </ModalContent>
+      </Modal>
+      <div>
+        <Table size="md">
+          <Thead>
+            <Tr>
+              <Th>Suffix</Th>
+              <Th>URL</Th>
+              <Th isNumeric>
+                <IconButton
+                  aria-label="Add New"
+                  icon={<FaPlus />}
+                  onClick={onOpen}
+                />
+              </Th>
+            </Tr>
+          </Thead>
+          <Tbody>
+            {urls.map((e, i) => (
+              <Tr key={i}>
+                <Td style={{ maxWidth: "10px" }}>{e.suffix}</Td>
+                <Td style={{ maxWidth: "30vw" }}>{e.url}</Td>
+                <Td isNumeric>
+                  <Menu>
+                    <MenuButton
+                      as={IconButton}
+                      aria-label="More Actions"
+                      icon={<FaEllipsisH />}
+                      variant=""
+                    />
+                    <MenuList>
+                      <MenuItem
+                        icon={<FaTrash />}
+                        onClick={() => deleteURL(e.suffix)}
+                      >
+                        Delete
+                      </MenuItem>
+                    </MenuList>
+                  </Menu>
+                </Td>
+              </Tr>
+            ))}
+          </Tbody>
+          <Tfoot>
+            <Tr>
+              <Th />
+              <Th />
+              <Th isNumeric />
+            </Tr>
+          </Tfoot>
+        </Table>
       </div>
-      {adminErrorMsg !== `` ? (
-        <div className={style.errorMsg}>
-          <FaExclamationCircle />
-          <p>{adminErrorMsg}</p>
-        </div>
-      ) : (
-        <></>
-      )}
-      <button onClick={createUrl}>Create</button>
-      <div className={style.statusOverlay}></div>
-    </div>
+    </>
   );
 }
