--- app/src/pages/url/style.module.css
@@ -1,36 +1,0 @@
-/* 
-
-.shortUrlDiv {
-  display: flex;
-  flex-direction: row;
-}
-
-.shortUrlDiv p {
-  margin-top: 15px;
-}
-
-.shortUrlDiv input {
-  width: 100%;
-  margin-left: 5px;
-}
-
-.domain {
-  font-size: 18px;
-}
-
-.errorMsg {
-  background-color: #ffd4d4;
-  padding: 20px;
-  border-radius: 10px;
-  display: flex;
-  place-items: center;
-}
-
-.errorMsg p {
-  margin-left: 10px;
-}
-
-.smallText {
-  font-size: 12px;
-  margin: 10px 0;
-} */
